requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64

pyyaml~=6.0.1
pandas==2.0.3
ezomero~=2.0.0
watchdog~=3.0.0
openpyxl~=3.1.2
mysql.connector~=2.2.9